requirements.txt
anthropic
boto3
botocore
bcrypt
celery
chromadb
eventlet
einops
Flask
Flask_Cors
Flask_Migrate
Flask_SocketIO
flask_sqlalchemy
Flask-JWT-Extended
Flask-RESTful
Flask-Mail
flask-smorest
flask-redis
flask-marshmallow
google-cloud-aiplatform
google-cloud-discoveryengine
google-search-results
huggingface_hub
jinja2
langchain
langchainhub
langchain-openai
langchain-anthropic
langchain-aws
langchain-community
langchain-google-genai
marshmallow-sqlalchemy
mysql-connector
numexpr
pandas
SQLAlchemy
pymysql
python-dotenv
requests
transformers
tavily-python
text-generation
Werkzeug
watchtower
openai
tiktoken
pypdf
sentencepiece
sentence_transformers
geoip2
geoip2
geocoder
